requirements.txt
h5py==2.8.0
numpy==1.14.5
opencv-python==3.4.3.18
pandas==0.23.4
tensorboard==1.10.0
tensorflow-gpu==1.10.0
ujson==1.35